refactor file names, remove go routine assersion, http routines are not killed on stop
