+++ service/stop.go
@@ -1,0 +1,7 @@
+package service
+
+import "context"
+
+type Stopper interface {
+	stop(context.Context)
+}
